Moving functions into a library session 1.
