+++ .gitignore
@@ -5,3 +5,4 @@
 /PhotoResizer/bin
 /PhotoResizer/obj
 /PhotoResizeLib/obj/Release
+/PhotoResizeLib/bin/Release
